@@ -1,8 +1,15 @@
-module github.com/boumenot/gocover-cobertura
+module github.com/franchb/gocover-cobertura
 
-go 1.14
+go 1.21
 
 require (
-	github.com/stretchr/testify v1.8.2
-	golang.org/x/tools v0.8.0
+	github.com/stretchr/testify v1.9.0
+	golang.org/x/tools v0.18.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/mod v0.15.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
